extracted parts of page into their own components
